fix(tests): Make the unit tests clang-analyzer clean

+++ tests/check_securechannel.c
@@ -281,7 +281,7 @@ START_TEST(SecureChannel_sendAsymmetricOPNMessage_sentDataIsValid) {
     for(size_t i = 0; i <= paddingSize; ++i) {
         ck_assert_msg(sentData.data[offset + i] == paddingByte,
                       "Expected padding byte %i to be %i but got value %i",
-                      i, paddingByte, sentData.data[offset + i]);
+                      (int)i, paddingByte, sentData.data[offset + i]);
     }
 
     ck_assert_msg(sentData.data[offset + paddingSize + 1] == '*', "Expected first byte of signature");
@@ -356,9 +356,7 @@ START_TEST(Securechannel_sendAsymmetricOPNMessage_extraPaddingPresentWhenKeyLarg
     for(size_t i = 0; i <= paddingSize; ++i) {
         ck_assert_msg(sentData.data[offset + i] == paddingByte,
                       "Expected padding byte %i to be %i but got value %i",
-                      i,
-                      paddingByte,
-                      sentData.data[offset + i]);
+                      (int)i, paddingByte, sentData.data[offset + i]);
     }
 
     ck_assert_msg(sentData.data[offset + paddingSize + 1] == extraPaddingByte,
